requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.5
drf-writable-nested==0.6.2
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
whitenoise==5.2.0
